+++ networkconfig/ssv.go
@@ -50,7 +50,7 @@ type SSVConfig struct {
 	// We currently support only MaxF=4, it should be changed only for experimental testing.
 	// IMPORTANT: While it's possible to change it, the codebase is not adapted to the change yet,
 	// so it doesn't work out of the box. The support will be added gradually.
-	MaxF int
+	MaxF uint8
 	// TotalEthereumValidators value needs to be maintained — consider getting it from external API
 	// with default or per-network value(s) as fallback
 	TotalEthereumValidators int
@@ -74,7 +74,7 @@ type marshaledConfig struct {
 	RegistryContractAddr    ethcommon.Address `json:"RegistryContractAddr,omitempty" yaml:"RegistryContractAddr,omitempty"`
 	Bootnodes               []string          `json:"Bootnodes,omitempty" yaml:"Bootnodes,omitempty"`
 	DiscoveryProtocolID     hexutil.Bytes     `json:"DiscoveryProtocolID,omitempty" yaml:"DiscoveryProtocolID,omitempty"`
-	MaxF                    int               `json:"MaxF,omitempty" yaml:"MaxF,omitempty"`
+	MaxF                    uint8             `json:"MaxF,omitempty" yaml:"MaxF,omitempty"`
 	TotalEthereumValidators int               `json:"TotalEthereumValidators,omitempty" yaml:"TotalEthereumValidators,omitempty"`
 	GasLimit36Epoch         phase0.Epoch      `json:"GasLimit36Epoch,omitempty" yaml:"GasLimit36Epoch,omitempty"`
 }
@@ -155,10 +155,10 @@ func (s *SSVConfig) MaxOperators() int {
 		maxF = defaultMaxF
 	}
 
-	return s.calcOperatorCount(maxF)
+	return int(s.calcOperatorCount(maxF))
 }
 
-func (s *SSVConfig) calcOperatorCount(f int) int {
+func (s *SSVConfig) calcOperatorCount(f uint8) uint8 {
 	// We heavily rely on this formula in the codebase, however, this function is made flexible intentionally
 	// to allow experiments on local testnet as there might exist better formulas theoretically
 	// (e.g. https://www.anza.xyz/blog/alpenglow-a-new-consensus-for-solana)
